[TEST] timing the vulkan render device's window_resize reaction
It appears that the `fpGetPhysicalDeviceSurfaceCapabilitiesKHR` calls take ~210ms and the `fpCreateSwapchainKHR` take ~390ms, at least on Debian bullseye with proprietary Nvidia 440.100.

That said, it appears that queueing the resizes until we start rendering a new frame doesn't fix the issue. It does help reduce time wasted on unnecessary swapchain reallocations though. Still, the reason the window doesn't get rendered is that by the time `VulkanContext::swap_buffers` is called, the vulkan is already returning `VK_ERROR_OUT_OF_DATE_KHR`, so that doesn't help much.