Add external editor presets
Choose preset to set exec_path and exec_flags to values for the chosen editor. Optionally change the exec_path to a non-empty path to use it as the editor path, while keeping the exec_flags preset for the selected editor.

Choose "Custom" (default) as the preset to use the old behavior.